backend/pyproject.toml
[project]
name = "billiken-blueprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "exceptiongroup>=1.3.0",
    "fastapi[standard]>=0.121.0",
    "google-genai>=1.47.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest-asyncio>=1.2.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
